backend/requirements.txt
# CortexDocs ∞ — Backend Dependencies
# Python 3.11+

# ─── Web Framework ─────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
starlette==0.38.6

# ─── Database ──────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.9

# ─── Validation & Settings ─────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ─── ML Models ─────────────────────────────────────
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
numpy==1.26.4
torch>=2.0.0

# ─── PDF Processing ───────────────────────────────
PyMuPDF==1.24.11

# ─── Resilience ────────────────────────────────────
tenacity==9.0.0

# ─── Observability ─────────────────────────────────
structlog==24.4.0

# ─── Monitoring ────────────────────────────────────
psutil==6.0.0

# ─── Testing ──────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
aiosqlite==0.20.0
